separate twirl dependent code into own package log the time it takes to deliver a page
